The structure is done for the laptop side.  Now the code needs to be filled in
