+++ Laptop/io.py
@@ -1,0 +1,8 @@
+# we need to import some sort of joystick reading library
+
+stick_x 0.0
+stick_y = 0.0
+
+def update_joysticks():
+    # update stick_x and stick_y vars
+    pass
